+++ spec/features/staff/users_can_view_budgets_on_an_activity_page_spec.rb
@@ -168,11 +168,7 @@ RSpec.feature "Users can view budgets on an activity page" do
   end
 
   def budget_information_is_shown_on_page(budget_presenter)
-    expect(page).to have_content(budget_presenter.budget_type)
-    expect(page).to have_content(budget_presenter.status)
-    expect(page).to have_content(budget_presenter.period_start_date)
-    expect(page).to have_content(budget_presenter.period_end_date)
-    expect(page).to have_content(budget_presenter.currency)
+    expect(page).to have_content(budget_presenter.financial_year)
     expect(page).to have_content(budget_presenter.value)
   end
 end